tools/pdf-to-dxf-ai/requirements.txt
httpx>=0.27.0
ezdxf>=1.0.3
PyMuPDF>=1.23.0
customtkinter>=5.2.0
Pillow>=10.0.0
matplotlib>=3.7.0
